Set default stop signal and insert tini init system
This change installs the tini init system to take over as PID 1 - tini will fulfill the responsibility of propagating the incoming signal to all child processes.

We also set the STOPSIGNAL in the Dockerfile - This sets the signal that docker will send to the container when asked to perform a generic container stop. Currently wings can only send a KILL or a generic stop